modified:   src/headers/app.helper.h 	modified:   src/helper.c 	modified:   src/process.c

+++ src/headers/app.helper.h
@@ -6,9 +6,9 @@ long int    stringLengthHelper              (const char * s);
 void        stringPrintHelper               (const char * s);
 void        integerPrintHelper              (int n);
 void        unsignedIntegerPrintHelper      (unsigned int n);
-void	    addressPrintHelper              (unsigned long long n);
-void        hexadecimalPrintHelper          (unsigned long long n);
-void        hexadecimalCapitalPrintHelper   (unsigned long long n);
+void	    addressPrintHelper              (unsigned int n);
+void        hexadecimalPrintHelper          (unsigned int n);
+void        hexadecimalCapitalPrintHelper   (unsigned int n);
 void	    octalPrintHelper                (unsigned int n);
 
 #endif //MY_PRINTF_APP_HELPER
